requirements.txt
# Bronze Tier Foundation - Python Dependencies
# Python 3.13+ required

# File system monitoring
watchdog>=4.0.0

# Environment variable management
python-dotenv>=1.0.0

# YAML parsing
pyyaml>=6.0

# Markdown frontmatter parsing
python-frontmatter>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Gmail watcher (optional)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
